+++ backend/api/controllers/psa.ts
@@ -1,4 +1,4 @@
-import * as PsaDAL from "../../dao/psa";
+import * as PsaDAL from "../../dal/psa";
 import { MonkeyResponse } from "../../utils/monkey-response";
 
 export async function getPsas(
