+++ backend/dal/bot.ts
