+++ backend/dal/new-quotes.ts
