+++ backend/dal/quote-ratings.ts
