+++ backend/dal/result.ts
